+++ src/pymatgen/analysis/compatibility/correction_calculator.py
@@ -15,8 +15,8 @@ from ruamel import yaml
 from scipy.optimize import curve_fit
 
 from pymatgen.analysis.reaction_calculator import ComputedReaction
-from pymatgen.analysis.structure_analyzer import sulfide_type
 from pymatgen.core import Composition, Element
+from pymatgen.core.structure_analyzer import sulfide_type
 
 logger = logging.getLogger(__name__)