+++ src/pymatgen/analysis/pourbaix_diagram.py
@@ -19,12 +19,12 @@ from monty.json import MontyDecoder, MSONable
 from scipy.spatial import ConvexHull, HalfspaceIntersection
 from scipy.special import comb
 
+from pymatgen.analysis.compatibility import MU_H2O
 from pymatgen.analysis.phase_diagram import PDEntry, PhaseDiagram
 from pymatgen.analysis.reaction_calculator import Reaction, ReactionError
 from pymatgen.core import Composition, Element
 from pymatgen.core.entries import ComputedEntry
 from pymatgen.core.ion import Ion
-from pymatgen.entries.compatibility import MU_H2O
 from pymatgen.util.coord import Simplex
 from pymatgen.util.due import Doi, due
 from pymatgen.util.plotting import pretty_plot