+++ src/pymatgen/entries/mixing_scheme.py
@@ -13,16 +13,16 @@ from itertools import groupby
 import numpy as np
 import pandas as pd
 
-from pymatgen.analysis.phase_diagram import PhaseDiagram
-from pymatgen.analysis.structure_matcher import StructureMatcher
-from pymatgen.core.entries import ComputedStructureEntry, ConstantEnergyAdjustment
-from pymatgen.entries.compatibility import (
+from pymatgen.analysis.compatibility import (
     Compatibility,
     CompatibilityError,
     MaterialsProject2020Compatibility,
     TypeVarAnyEntry,
 )
-from pymatgen.entries.entry_tools import EntrySet
+from pymatgen.analysis.compatibility.entry_tools import EntrySet
+from pymatgen.analysis.phase_diagram import PhaseDiagram
+from pymatgen.core.entries import ComputedStructureEntry, ConstantEnergyAdjustment
+from pymatgen.core.structure_matcher import StructureMatcher
 
 MODULE_DIR = os.path.dirname(os.path.abspath(__file__))