+++ src/pymatgen/vis/structure_chemview.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import numpy as np
 
-from pymatgen.analysis.molecule_structure_comparator import CovalentRadius
+from pymatgen.core.molecule_structure_comparator import CovalentRadius
 from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 
 try: