+++ tests/analysis/test_graphs.py
@@ -12,8 +12,10 @@ import pytest
 from monty.serialization import loadfn
 from pytest import approx
 
-from pymatgen.analysis.graphs import MoleculeGraph, MolGraphSplitError, PeriodicSite, StructureGraph
-from pymatgen.analysis.local_env import (
+from pymatgen.command_line.critic2_caller import Critic2Analysis
+from pymatgen.core import Lattice, Molecule, Site, Structure
+from pymatgen.core.graphs import MoleculeGraph, MolGraphSplitError, PeriodicSite, StructureGraph
+from pymatgen.core.local_env import (
     CovalentBondNN,
     CutOffDictNN,
     MinimumDistanceNN,
@@ -21,8 +23,6 @@ from pymatgen.analysis.local_env import (
     OpenBabelNN,
     VoronoiNN,
 )
-from pymatgen.command_line.critic2_caller import Critic2Analysis
-from pymatgen.core import Lattice, Molecule, Site, Structure
 from pymatgen.core.structure import FunctionalGroups
 from pymatgen.util.testing import TEST_FILES_DIR, MatSciTest