+++ tests/analysis/test_local_env.py
@@ -9,8 +9,9 @@ import pytest
 from numpy.testing import assert_allclose
 from pytest import approx
 
-from pymatgen.analysis.graphs import MoleculeGraph, StructureGraph
-from pymatgen.analysis.local_env import (
+from pymatgen.core import Element, Lattice, Molecule, Structure
+from pymatgen.core.graphs import MoleculeGraph, StructureGraph
+from pymatgen.core.local_env import (
     CN_OPT_PARAMS,
     DEFAULT_OP_PARAMS,
     BrunnerNNReal,
@@ -40,7 +41,6 @@ from pymatgen.analysis.local_env import (
     site_is_of_motif_type,
     solid_angle,
 )
-from pymatgen.core import Element, Lattice, Molecule, Structure
 from pymatgen.util.testing import TEST_FILES_DIR, MatSciTest
 
 TEST_DIR = f"{TEST_FILES_DIR}/core/local_env/fragmenter_files"