+++ tests/analysis/test_phase_diagram.py
@@ -14,6 +14,7 @@ from monty.serialization import dumpfn, loadfn
 from numpy.testing import assert_allclose
 from pytest import approx
 
+from pymatgen.analysis.compatibility.entry_tools import EntrySet
 from pymatgen.analysis.phase_diagram import (
     CompoundPhaseDiagram,
     GrandPotentialPhaseDiagram,
@@ -30,7 +31,6 @@ from pymatgen.analysis.phase_diagram import (
 )
 from pymatgen.core import Composition, DummySpecies, Element
 from pymatgen.core.entries import ComputedEntry
-from pymatgen.entries.entry_tools import EntrySet
 from pymatgen.util.testing import TEST_FILES_DIR, MatSciTest
 
 TEST_DIR = f"{TEST_FILES_DIR}/analysis"